requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
jwt==1.3.1
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.1
